requirements.txt
ipfabric
loguru
python3-dotenv
typer
